requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# AI Model APIs
anthropic>=0.25.0
openai>=1.0.0
google-generativeai>=0.8.0

# Optional dependencies for enhanced functionality
# numpy>=1.24.0  # For statistical analysis
# pandas>=2.0.0  # For data manipulation and export
# matplotlib>=3.7.0  # For visualization